requirements.txt
pandas==2.0.3
numpy==1.26.0
matplotlib==3.8.0
streamlit==1.37.0
sentence-transformers==2.2.2
transformers==4.41.0
huggingface-hub==0.13.4
openai==1.29.0
python-dotenv==1.0.1
faiss-cpu==1.7.4
